+++ vtm-android-example/src/org/oscim/android/test/Samples.java
@@ -80,12 +80,12 @@ public class Samples extends Activity {
         setContentView(R.layout.activity_samples);
         LinearLayout linearLayout = findViewById(R.id.samples);
         linearLayout.addView(createButton(SimpleMapActivity.class));
-        linearLayout.addView(createButton(MapsforgeMapActivity.class));
-        /*linearLayout.addView(createButton(MapzenMvtMapActivity.class));
-        linearLayout.addView(createButton(MapzenGeojsonMapActivity.class));*/
-        linearLayout.addView(createButton(OpenMapTilesMvtMapActivity.class));
-        linearLayout.addView(createButton(OpenMapTilesGeojsonMapActivity.class));
-        linearLayout.addView(createButton(GdxMapActivity.class));
+        linearLayout.addView(createButton(MapsforgeActivity.class));
+        /*linearLayout.addView(createButton(MapzenMvtActivity.class));
+        linearLayout.addView(createButton(MapzenGeojsonActivity.class));*/
+        linearLayout.addView(createButton(OpenMapTilesMvtActivity.class));
+        linearLayout.addView(createButton(OpenMapTilesGeojsonActivity.class));
+        linearLayout.addView(createButton(GdxActivity.class));
 
         linearLayout.addView(createLabel("Features"));
         linearLayout.addView(createButton(null, "GraphHopper Routing", new View.OnClickListener() {
@@ -104,7 +104,7 @@ public class Samples extends Activity {
         linearLayout.addView(createButton(POTTextureActivity.class));
 
         linearLayout.addView(createLabel("Raster Maps"));
-        linearLayout.addView(createButton(BitmapTileMapActivity.class));
+        linearLayout.addView(createButton(BitmapTileActivity.class));
 
         linearLayout.addView(createLabel("Overlays"));
         linearLayout.addView(createButton(MarkerOverlayActivity.class));
@@ -114,7 +114,7 @@ public class Samples extends Activity {
         linearLayout.addView(createButton(ClusterMarkerOverlayActivity.class));
         linearLayout.addView(createButton(PathOverlayActivity.class));
         linearLayout.addView(createButton(LineTexActivity.class));
-        linearLayout.addView(createButton(VectorLayerMapActivity.class));
+        linearLayout.addView(createButton(VectorLayerActivity.class));
         linearLayout.addView(createButton(LocationActivity.class));
 
         linearLayout.addView(createLabel("User Interaction"));
@@ -122,12 +122,12 @@ public class Samples extends Activity {
         linearLayout.addView(createButton(LayerGroupActivity.class));
 
         linearLayout.addView(createLabel("Dual Map Views"));
-        linearLayout.addView(createButton(MultiMapActivity.class));
+        linearLayout.addView(createButton(MultiMapViewActivity.class));
 
         linearLayout.addView(createLabel("Experiments"));
         linearLayout.addView(createButton(ReverseGeocodeActivity.class));
-        linearLayout.addView(createButton(OSciMapS3DBMapActivity.class));
+        linearLayout.addView(createButton(OSciMapS3DBActivity.class));
         linearLayout.addView(createButton(ThemeStylerActivity.class));
-        linearLayout.addView(createButton(JeoIndoorMapActivity.class));
+        linearLayout.addView(createButton(JeoIndoorActivity.class));
     }
 }