+++ vtm-playground/src/org/oscim/test/BitmapTileTest.java
@@ -24,7 +24,7 @@ import org.oscim.layers.tile.bitmap.BitmapTileLayer;
 import org.oscim.renderer.MapRenderer;
 import org.oscim.tiling.source.bitmap.DefaultSources;
 
-public class BitmapLayerTest extends GdxMapImpl {
+public class BitmapTileTest extends GdxMapImpl {
 
     private BitmapTileLayer mLayer = null;
     private BitmapTileLayer mShaded = null;
@@ -73,6 +73,6 @@ public class BitmapLayerTest extends GdxMapImpl {
 
     public static void main(String[] args) {
         GdxMapApp.init();
-        GdxMapApp.run(new BitmapLayerTest(), null, 256);
+        GdxMapApp.run(new BitmapTileTest(), null, 256);
     }
 }