+++ docs/Changelog.md
@@ -3,8 +3,8 @@
 ## New since 0.9.2
 
 - S3DB layer [#475](https://github.com/mapsforge/vtm/pull/475)
-- vtm-mvt module with MVT tile decoder [#481](https://github.com/mapsforge/vtm/issues/481)
-- OpenMapTiles MVT vector tiles [#481](https://github.com/mapsforge/vtm/issues/481)
+- vtm-mvt module with MVT tile decoder [#481](https://github.com/mapsforge/vtm/pull/481)
+- OpenMapTiles MVT vector tiles [#481](https://github.com/mapsforge/vtm/pull/481)
 - OpenMapTiles GeoJSON vector tiles [#385](https://github.com/mapsforge/vtm/issues/385)
 - Render theme fallback internal resources [#477](https://github.com/mapsforge/vtm/issues/477)
 - Many other minor improvements and bug fixes